Use new loki API
Switch to the new loki API As the old prom API is
[deprecated](https://github.com/grafana/loki/blob/master/docs/api.md#get-apipromquery).
This uses the client from the logcli package so we don't need to worry
about marshaling/unmarshaling anymore. Regexp filtering via parameter is
not possible anymore using the new API so that field has been removed.
Filtering should be done directly in the query with logql.

+++ pkg/config/config.go
@@ -28,7 +28,7 @@ type Config struct {
 	Queries []struct {
 		Name      string        `yaml:"name"`
 		Query     string        `yaml:"query"`
-		Limit     int64         `yaml:"limit"`
+		Limit     int           `yaml:"limit"`
 		Start     time.Duration `yaml:"start"`
 		End       time.Duration `yaml:"end"`
 		Direction string        `yaml:"direction"`
